requirements.txt
keras_applications>=1.0.7,<=1.0.8
